+++ changelog.md
@@ -3,6 +3,7 @@
 #### Version 1.9.1 - Unreleased
 - A custom desert map option. - NicEastvillage
 - Mercy rule is now optional and adjustable. - NicEastvillage
+- The match is automatically stopped once mercy is hit. - NicEastvillage
 - Fix `show_mercy.json` not being written to when mercy was hit. - NicEastvillage
 - Fix `show_vs.json` not being written to on match startup and match begin. - NicEastvillage
 - Fix path used by "Load BotPack" now that RLBot v5 has settled. - NicEastvillage